+++ lib/GaletteObjectsLend/Entity/LendCategory.php
@@ -39,13 +39,13 @@
  * @since     Available since 0.7
  */
 
-namespace GaletteObjectsLend\Repository;
+namespace GaletteObjectsLend\Entity;
 
 use Analog\Analog;
 use Galette\Core\Db;
 use Galette\Core\Plugins;
 use \Zend\Db\Sql\Predicate;
-use GaletteObjectsLend\Repository\CategoryPicture;
+use GaletteObjectsLend\Entity\CategoryPicture;
 
 class LendCategory
 {
