+++ lib/GaletteObjectsLend/Entity/LendPDF.php
@@ -39,7 +39,7 @@
  * @since     Available since 0.7
  */
 
-namespace GaletteObjectsLend\Repository;
+namespace GaletteObjectsLend\Entity;
 
 use Galette\Core\Preferences as CorePreferences;
