+++ lib/GaletteObjectsLend/Entity/ObjectPicture.php
@@ -3,7 +3,7 @@
 /* vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4: */
 
 /**
- * Picture of a category
+ * Picture of an object
  *
  * PHP version 5
  *
@@ -38,15 +38,15 @@
  * @since     Available since 0.7
  */
 
-namespace GaletteObjectsLend\Repository;
+namespace GaletteObjectsLend\Entity;
 
 use Analog\Analog;
 use Galette\Core\Plugins;
 
-class CategoryPicture extends Picture
+class ObjectPicture extends Picture
 {
-    const PK = 'category_id';
-    const TABLE = 'categories_pictures';
+    const PK = 'object_id';
+    const TABLE = 'pictures';
 
     /**
      * Default constructor.
@@ -56,8 +56,7 @@ class CategoryPicture extends Picture
      */
     public function __construct(Plugins $plugins, $objectid = '')
     {
-        $this->store_path = GALETTE_PHOTOS_PATH . 'objectslend/categories/';
-
+        $this->store_path = GALETTE_PHOTOS_PATH . 'objectslend/objects/';
         parent::__construct($plugins, $objectid);
     }
 }
