+++ lib/GaletteObjectsLend/IO/PdfObjects.php
@@ -41,8 +41,8 @@ use Galette\Core\Preferences;
 use Galette\Core\Login;
 use Analog\Analog;
 use GaletteObjectsLend\Filters\ObjectsList;
-use GaletteObjectsLend\Repository\LendCategory;
-use GaletteObjectsLend\Repository\Preferences as LendPreferences;
+use GaletteObjectsLend\Entity\LendCategory;
+use GaletteObjectsLend\Entity\Preferences as LendPreferences;
 
 /**
  * Object labels PDF
@@ -109,7 +109,7 @@ class PdfObjects extends Pdf
      *
      * @return void
      */
-    public function Header()
+    public function header()
     {
         $this->SetFont(Pdf::FONT, 'B');
         $x = $this->getX();
@@ -169,7 +169,10 @@ class PdfObjects extends Pdf
 
         $this->Cell($w_checkbox, 0, $this->stretchHead('', $w_checkbox), 1, 0, 'C', 1);
         $this->Cell($w_name, 0, $this->stretchHead(_T("Name", "objectslend"), $w_name), 1, 0, 'C', 1);
-        $this->Cell($w_description, 0, $this->stretchHead(_T("Description", "objectslend"), $w_description), 1, 0, 'C', 1);
+        $this->Cell($w_description, 0, $this->stretchHead(
+            _T("Description", "objectslend"),
+            $w_description
+        ), 1, 0, 'C', 1);
         $this->Cell($w_serial, 0, $this->stretchHead(_T("Serial", "objectslend"), $w_serial), 1, 0, 'C', 1);
         $this->Cell($w_price, 0, $this->stretchHead(_T("Price", "objectslend"), $w_price), 1, 0, 'C', 1);
         $this->Cell($w_price, 0, $this->stretchHead(_T("Borrow price", "objectslend"), $w_price), 1, 0, 'C', 1);
@@ -188,10 +191,11 @@ class PdfObjects extends Pdf
         $row = 0;
 
         foreach ($objects as $object) {
-			$rents = $object->rents;
+            $rents = $object->rents;
 
-            if ($this->lendsprefs->{LendPreferences::PARAM_VIEW_CATEGORY} && $current_category !== $object->category_id) {
-                $this->SetFont('', 'B');
+            if ($this->lendsprefs->{LendPreferences::PARAM_VIEW_CATEGORY} &&
+                $current_category !== $object->category_id) {
+                    $this->SetFont('', 'B');
 
                 if (($this->login->isAdmin() || $this->login->isStaff()) && $sum_price > 0) {
                     $width = $w_checkbox + $w_name + $w_description + $w_serial + $w_price;
@@ -201,7 +205,6 @@ class PdfObjects extends Pdf
                 }
 
                 if ($object->category_name) {
-
                     $text = str_replace(
                         '%category',
                         $object->category_name,
@@ -211,8 +214,8 @@ class PdfObjects extends Pdf
                     $text = _T("No category");
                 }
 
-                $this->Cell(0, 0, $text, 0, 1, 'C');
-                $this->SetFont('');
+                    $this->Cell(0, 0, $text, 0, 1, 'C');
+                    $this->SetFont('');
             }
 
             if ($row++ % 2 == 0) {
@@ -221,7 +224,7 @@ class PdfObjects extends Pdf
                 $this->SetFillColor(255, 214, 135);
             }
 
-			$fill=!$object->is_home_location;
+            $fill=!$object->is_home_location;
 
             $this->Cell($w_checkbox, 0, '□', 'B', 0, 'L', $fill);
             $this->Cell($w_name, 0, $this->cut($object->name, $w_name), 'B', 0, 'L', $fill);
@@ -233,7 +236,8 @@ class PdfObjects extends Pdf
             $this->Cell($w_weight, 0, $this->cut($object->weight, $w_weight), 'B', 0, 'R', $fill);
             $this->Cell($w_status, 0, $this->cut($object->status_text, $w_status), 'B', 0, 'L', $fill);
             $this->Cell($w_date, 0, $this->cut($object->date_begin_short, $w_date), 'B', 0, 'L', $fill);
-            $this->Cell($w_adherent, 0, $this->cut($object->nom_adh . ' ' . $object->prenom_adh, $w_adherent), 'B', 0, 'L', $fill);
+            $this->Cell($w_adherent, 0, $this->cut($object->nom_adh . ' ' .
+                $object->prenom_adh, $w_adherent), 'B', 0, 'L', $fill);
             $this->Cell($w_date, 0, $this->cut($object->date_forecast_short, $w_date), 'B', 1, 'L', $fill);
 
             if ($this->login->isAdmin() || $this->login->isStaff()) {
@@ -248,7 +252,8 @@ class PdfObjects extends Pdf
             $this->Ln();
             $this->Ln();
 
-            $this->Cell($width, 0, _T("Sum:", "objectslend") . ' ' . number_format($grant_total, 2, ',', ''), '', 0, 'R');
+            $this->Cell($width, 0, _T("Sum:", "objectslend") . ' ' .
+                number_format($grant_total, 2, ',', ''), '', 0, 'R');
             $this->Ln();
         }
 
@@ -271,7 +276,7 @@ class PdfObjects extends Pdf
      */
     public function drawList1($objects)
     {
-		$this->Open();
+        $this->Open();
         $this->AddPage();
 
         $this->Ln(10); //for Header
@@ -285,38 +290,37 @@ class PdfObjects extends Pdf
 
 
 
-		$col_begin = 33;
-		$col_end = 33;
-		$col_status = 30;
-		$col_home = 20;
-		$col_adh = 30;
-		$col_comments = 128;
+        $col_begin = 33;
+        $col_end = 33;
+        $col_status = 30;
+        $col_home = 20;
+        $col_adh = 30;
+        $col_comments = 128;
 
-		$this->SetFont(Pdf::FONT, 'B', 10);
-		$this->Cell(0, 0, _T("History of objects loans", "objectslend"), 0, 1, 'C');
-		$this->Ln();
-		$this->Cell($col_begin, 0,$this->cut(_T("Begin", "objectslend"), $col_begin), 'B');
-		$this->Cell($col_end, 0,$this->cut(_T("End", "objectslend"), $col_end), 'B');
-		$this->Cell($col_status, 0,$this->cut(_T("Status", "objectslend"), $col_status), 'B');
-		$this->Cell($col_home, 0,$this->cut(_T("On site", "objectslend"), $col_home), 'B');
-		$this->Cell($col_adh, 0,$this->cut(_T("Member", "objectslend"), $col_adh), 'B');
-		$this->Cell($col_comments, 0,$this->cut(_T("Comments", "objectslend"), $col_comments), 'B');
-		$this->Ln();
-		$this->SetFont(Pdf::FONT, '', 9);
+        $this->SetFont(Pdf::FONT, 'B', 10);
+        $this->Cell(0, 0, _T("History of objects loans", "objectslend"), 0, 1, 'C');
+        $this->Ln();
+        $this->Cell($col_begin, 0, $this->cut(_T("Begin", "objectslend"), $col_begin), 'B');
+        $this->Cell($col_end, 0, $this->cut(_T("End", "objectslend"), $col_end), 'B');
+        $this->Cell($col_status, 0, $this->cut(_T("Status", "objectslend"), $col_status), 'B');
+        $this->Cell($col_home, 0, $this->cut(_T("On site", "objectslend"), $col_home), 'B');
+        $this->Cell($col_adh, 0, $this->cut(_T("Member", "objectslend"), $col_adh), 'B');
+        $this->Cell($col_comments, 0, $this->cut(_T("Comments", "objectslend"), $col_comments), 'B');
+        $this->Ln();
+        $this->SetFont(Pdf::FONT, '', 9);
 
-		foreach ($objects as $object) {
-			 $rents = $object->rents;
-			foreach ($rents as $rt) {
-				$this->Cell($col_begin, 0,$this->cut($rt->date_begin, $col_begin), 'B');
-				$this->Cell($col_end, 0,$this->cut($rt->date_end, $col_end), 'B');
-				$this->Cell($col_status, 0,$this->cut($rt->status_text, $col_status), 'B');
-				$this->Cell($col_home, 0, $rt->is_home_location ? '    X' : '', 'B');
-				$this->Cell($col_adh, 0,$this->cut($rt->nom_adh . ' ' . $rt->prenom_adh, $col_adh), 'B');
-				$this->Cell($col_comments, 0,$this->cut($rt->comments, $col_comments), 'B');
-				$this->Ln();
-			}
-		}
-
+        foreach ($objects as $object) {
+             $rents = $object->rents;
+            foreach ($rents as $rt) {
+                $this->Cell($col_begin, 0, $this->cut($rt->date_begin, $col_begin), 'B');
+                $this->Cell($col_end, 0, $this->cut($rt->date_end, $col_end), 'B');
+                $this->Cell($col_status, 0, $this->cut($rt->status_text, $col_status), 'B');
+                $this->Cell($col_home, 0, $rt->is_home_location ? '    X' : '', 'B');
+                $this->Cell($col_adh, 0, $this->cut($rt->nom_adh . ' ' . $rt->prenom_adh, $col_adh), 'B');
+                $this->Cell($col_comments, 0, $this->cut($rt->comments, $col_comments), 'B');
+                $this->Ln();
+            }
+        }
     }
 
     /**
